refactor(editor): Move useToast into @n8n/composables behind a shim (no-changelog)
Move useToast composable and its tests from editor-ui into the
@n8n/composables package as part of the CAT-3686 frontend-modularization
effort.

Key changes:
- git mv useToast.ts and useToast.test.ts to @n8n/composables/src/
- Replace element-plus ElNotification import with a registration
  pattern (setNotify/NotifyFn) to avoid DTS generation OOM
- Inline route-based workflowId logic (previously useWorkflowId)
- Inline TOASTS_Z_INDEX constant (previously from useStyles)
- Repoint sanitizeHtml to @n8n/frontend-utils/htmlUtils
- Create deprecated re-export shim at old editor-ui path
- Update @n8n/composables package.json with required dependencies
