+++ packages/frontend/@n8n/composables/package.json
@@ -29,6 +29,10 @@
   },
   "dependencies": {
     "@n8n/api-types": "workspace:*",
+    "@n8n/frontend-constants": "workspace:*",
+    "@n8n/frontend-utils": "workspace:*",
+    "@n8n/i18n": "workspace:*",
+    "@n8n/stores": "workspace:*",
     "@vueuse/core": "catalog:frontend",
     "lodash": "catalog:",
     "n8n-workflow": "workspace:*",
@@ -41,12 +45,15 @@
     "@n8n/telemetry": "workspace:*",
     "@n8n/typescript-config": "workspace:*",
     "@n8n/vitest-config": "workspace:*",
+    "@pinia/testing": "^0.1.6",
+    "element-plus": "catalog:frontend",
     "@testing-library/vue": "catalog:frontend",
     "@testing-library/jest-dom": "catalog:frontend",
     "@testing-library/user-event": "catalog:frontend",
     "@types/lodash": "catalog:",
     "@vitejs/plugin-vue": "catalog:frontend",
     "@vue/tsconfig": "catalog:frontend",
+    "pinia": "catalog:frontend",
     "tsdown": "catalog:",
     "typescript": "catalog:",
     "vite": "catalog:",
